Changes in demo files.
